WIP: DO NOT MERGE
The goal of this patch is to set the v1trapaddress of PDUs sent via
internet protocols (?) (i.e. TCP, UDP) to the IP(v4??) address of the
outgoing interface. I _think_ this should be possible via some kind of
"get route" call, whether via 'ip route get' or RTNETLINK or similar.